+++ pkg/executor/executor.go
@@ -1,4 +1,4 @@
-package provisioning
+package executor
 
 // Executor is responsible for creating execution environment for given workload.
 // It returns a Task interface.
