+++ pkg/executor/local.go
@@ -1,4 +1,4 @@
-package provisioning
+package executor
 
 import (
 	"bytes"
