+++ pkg/executor/task.go
@@ -1,4 +1,4 @@
-package provisioning
+package executor
 
 // Task represents a process which can be stoped or monitored.
 type Task interface{
